Split up the big file and restructured to allow bitmap rendering in the future
