+++ Cargo.toml
@@ -47,6 +47,7 @@ serde_json = "1.0"
 similar = "2.7"
 thin-vec = { version = "0.2", features = ["serde"] }
 uom = { version = "0.36", features = ["use_serde", "usize", "isize"] }
+zeno = {version = "0.3.2" }
 
 [workspace.lints.rust]
 unexpected_cfgs = { level = "allow", check-cfg = [
