+++ rustyms/README.md
@@ -80,3 +80,4 @@ It has multiple features which allow you to slim it down if needed (all are enab
 * `rand` - allows the generation of random peptides.
 * `rayon` - enables parallel iterators using rayon, mostly for `imgt` but also in consecutive align.
 * `mzdata` - enables integration with [mzdata](https://github.com/mobiusklein/mzdata) which has more advanced raw file support.
+* `zeno` - enables the rendering to bitmaps for glycans
